+++ prettier.config.js
@@ -1,6 +1,7 @@
 import { createConfig } from '@alexlit/config-prettier';
 
 export default createConfig({
+  jsdoc: false,
   php: true,
   ruby: true,
   sql: true,