+++ scripts/preinstall.sh
@@ -2,6 +2,4 @@ source "$(dirname $0)"/../bash.config.sh
 
 echo -e "${YELLOW}Выполняю preinstal-скрипты... ${WHITE}"
 
-npx only-allow pnpm
-
 SAY_GOODBYE